fix(webgl): normalize unsigned integer operations
Use unsigned operands for mip-level shifting and normalize primitive-restart constants to uint64_t.



AI[100%] Human Fixed[0%] Human[0%] AI Adopted[100%]

+++ interfaces/kits/napi/graphic/webgl/src/webgl2_rendering_context_impl.cpp
@@ -880,7 +880,8 @@ GLenum WebGL2RenderingContextImpl::CheckTexImage3D(napi_env env, const TexImageA
     if (maxWidthHeight <= 0 || (imgArg.target == GL_TEXTURE_2D_ARRAY && maxArrayTextureLayers_ <= 0)) {
         return WebGLRenderingContextBase::INVALID_OPERATION;
     }
-    const GLint maxMipDimension = maxWidthHeight >> imgArg.level;
+    const GLint maxMipDimension = static_cast<GLint>(
+        static_cast<GLuint>(maxWidthHeight) >> static_cast<GLuint>(imgArg.level));
     const GLint maxDepth = imgArg.target == GL_TEXTURE_3D ? maxMipDimension : maxArrayTextureLayers_;
     if (imgArg.width > maxMipDimension || imgArg.height > maxMipDimension || imgArg.depth > maxDepth) {
         return WebGLRenderingContextBase::INVALID_VALUE;
